requirements.txt
jira~=3.5.2
python-gitlab~=4.1.1